Remove inline into cpp
