+++ simulator/runtime/rdobase.h
@@ -147,6 +147,4 @@ private:
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/rdobase-inl.h"
-
 #endif // _LIB_RUNTIME_BASE_H_
