+++ simulator/runtime/rdoframe.h
@@ -638,6 +638,4 @@ DECLARE_POINTER(RDOFRMFrame);
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/rdoframe-inl.h"
-
 #endif // _LIB_RUNTIME_FRAME_H_
